% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fars_functions.r
\name{fars_map_state}
\alias{fars_map_state}
\title{Plots the accidents on a US state map
The function accepts a state number and the year to plot the accidents on a single map
State number needs to be exixt in FARS data otherwise the fuction will return invalid state}
\usage{
fars_map_state(state.num, year)
}
\arguments{
\item{state.num}{The number of a state in the US as used in the FARS data
sets. Should be numeric or integer.}

\item{year}{The year of analysis (numeric or integer)}
}
\value{
a plot of the accidents based on the year and the state number
}
\description{
Plots the accidents on a US state map
The function accepts a state number and the year to plot the accidents on a single map
State number needs to be exixt in FARS data otherwise the fuction will return invalid state
}
\examples{
\dontrun{
fars_map_state(26, 2013)}
}
