% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fars_functions.r
\name{fars_read}
\alias{fars_read}
\title{Reads .csv file}
\usage{
fars_read(filename)
}
\arguments{
\item{filename}{as an input character vector}
}
\value{
the data frame dplyr::tbl_df() contains the whole data set or null if the file doesn't exist
}
\description{
This is a generic function reads the .csv file contains the data data from the US National Highway Traffic Safety Administration's
system \url{https://www.nhtsa.gov/Data/Fatality-Analysis-Reporting-System-(FARS)} and then create data frame with that data.
}
