% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fars_functions.r
\name{fars_read_years}
\alias{fars_read_years}
\title{Reads month and year from accident files}
\usage{
fars_read_years(years)
}
\arguments{
\item{years}{input numeric vector}
}
\value{
tibble or null if the year is invalid
}
\description{
The function accepts a vector or list of years and returns a list of data
frames with MONTH and year columns based on data in "accident_<year>.csv.bz2
files. The files need to be located in the working directory.
}
\examples{
\dontrun{
fars_read_years(2013:2015)
fars_read_years(list(2013, 2014))
}
}
