% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fars_functions.r
\name{fars_summarize_years}
\alias{fars_summarize_years}
\title{Counts number of accidents per month and year}
\usage{
fars_summarize_years(years)
}
\arguments{
\item{years}{a vector or list of years in numeric format}
}
\value{
a tibbol based on the months in rows and selected years in columns cotanining the number of accidents
}
\description{
Based on the list of years, the function calculates the number of accidents
in the US on a monthly basis. The accident files need to be in the working
directory, the years can be passed as a list or a vector.
}
\examples{
\dontrun{
fars_summarize_years(2013:2016)
}
}
