% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fars_functions.r
\name{make_filename}
\alias{make_filename}
\title{This function creats a filename
This function returns a STRING based on the year}
\usage{
make_filename(year)
}
\arguments{
\item{year}{as a numeric operator}
}
\value{
sprintf output STRING fromat based on the year passed as an parameter
}
\description{
This function creats a filename
This function returns a STRING based on the year
}
